Improve naming and delete old code
I've renamed the Nobt component to "HomeScreen" so it better reflects
what it actually is. In addition I deleted both, the sort and the filter
components as they are no longer needed.
BillItem is also no longer used.
