--- src/routes/App/components/BillItem/BillCardTextTheme.scss
@@ -1,3 +1,0 @@
-.cardText {
-  white-space: normal;
-}
