--- src/routes/App/components/BillItem/BillCardTitleTheme.scss
@@ -1,11 +1,0 @@
-.cardTitle {
-  padding-top: 1rem !important;
-}
-
-.title {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-@import "BillCardTitleTheme.shame";
