--- src/routes/App/components/BillItem/CardTheme.scss
@@ -1,7 +1,0 @@
-@import "~react-toolbox/lib/list/config";
-
-.card {
-  margin: $list-horizontal-padding;
-  margin-top: 0;
-  width: auto;
-}
