--- src/routes/App/components/BillItem/index.js
@@ -1,3 +1,0 @@
-import BillItem from "./BillItem";
-
-export default BillItem
