+++ src/routes/App/components/HomeScreen/HomeScreen.scss
@@ -1,24 +1,7 @@
-@import "~react-toolbox/lib/list/config";
-
-.nobt {
+.homeScreen {
   margin: auto;
   background-color: white;
   min-height: 100vh;
-}
-
-.cardListHeader {
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-
-  padding-top: $list-vertical-padding;
-  padding-right: $list-horizontal-padding;
-  padding-left: $list-horizontal-padding;
-}
-
-.addBillButton {
-  color: $color-accent-dark !important;
 }
 
 .overviewContainer{
