+++ src/routes/App/components/HomeScreen/index.js
@@ -1,0 +1,1 @@
+export default from "./HomeScreen.js"
