--- src/routes/App/routes/filter/index.js
@@ -1,10 +1,0 @@
-import FilterBillOverlay from "./components/FilterBillOverlay";
-import NobtContainer from "../../containers/NobtContainer";
-
-export default {
-  path: "changeFilter",
-  component: NobtContainer,
-  indexRoute: {
-    component: FilterBillOverlay
-  }
-}
