--- src/routes/App/routes/sort/components/SortBillOverlay.js
@@ -1,47 +1,0 @@
-import React from "react";
-import { connect } from "react-redux";
-import withNavigation from "components/hoc/withNavigation";
-import { ListItem } from "react-toolbox/lib/list";
-import { updateBillSortProperty } from "../../../modules/viewState/actions";
-import Dialog from "components/Dialog";
-import { getBillSortProperty } from "../../../modules/viewState/selectors";
-import HOList from "containers/HOList/HOList";
-import DialogTitle from "components/Dialog/DialogTitle";
-
-
-const SortBillOverlay = (props) => (
-
-  <Dialog>
-    <DialogTitle>Sort bills</DialogTitle>
-
-    <HOList
-      selectable
-      items={ [
-        {name: "Date", icon: "access_time"},
-        {name: "Amount", icon: "timeline"}
-      ]}
-      renderItem={ (config) => (
-        <ListItem
-          rightIcon={ config.name === props.currentSort ? "check_circle" : "radio_button_unchecked" }
-          key={config.name}
-          caption={config.name}
-          leftIcon={config.icon}
-          onClick={ () => {
-            props.updateBillSortProperty(config.name);
-            props.goBack()
-          } } />
-      )}>
-    </HOList>
-  </Dialog>
-);
-
-export default withNavigation(
-  connect(
-    (state) => ({
-      currentSort: getBillSortProperty(state)
-    }),
-    (dispatch) => ({
-      updateBillSortProperty: (property) => dispatch(updateBillSortProperty(property)),
-    })
-  )(SortBillOverlay)
-)
