--- src/routes/App/routes/sort/index.js
@@ -1,10 +1,0 @@
-import SortBillOverlay from "./components/SortBillOverlay";
-import NobtContainer from "../../containers/NobtContainer";
-
-export default {
-  path: "changeSort",
-  component: NobtContainer,
-  indexRoute: {
-    component: SortBillOverlay
-  }
-}
